+++ core/vgui/impl/mfc/vgui_mfc_dialog_impl.h
@@ -41,7 +41,7 @@ class vgui_mfc_dialog_impl : public CWnd,public vgui_dialog_impl
   ~vgui_mfc_dialog_impl();
 
   void* choice_field_widget(const char*, const vcl_vector<vcl_string>&, int&);
-  void* inline_tableau_widget(const vgui_tableau_sptr tab, float width, float height);
+  void* inline_tableau_widget(const vgui_tableau_sptr tab, unsigned width, unsigned height);
   //: Sets the modality of the dialog box.
   void modal(const bool);
   //: Display the dialog box.
